♻️ (#355) 아바타 그리드와 배경색 그리드 컴포넌트 분리
