+++ src/features/avatar-picker/components/AvatarsDrawer.tsx
@@ -1,4 +1,3 @@
-import { Avatar, AvatarFallback, AvatarImage } from '@/shared/components/shadcn/avatar';
 import {
   Drawer,
   DrawerContent,
@@ -7,17 +6,16 @@ import {
   DrawerTitle,
   DrawerTrigger,
 } from '@/shared/components/shadcn/drawer';
-import { Check, Pen, User } from 'lucide-react';
-import { cn } from '@/shared/lib/utils';
+import { Check, Pen } from 'lucide-react';
 import { AVATAR_BG_COLOR } from '@/features/avatar-picker/constants/avatarBgColor';
 import { useAvatarStore } from '@/features/avatar-picker/store/useAvatarStore';
-import { avatarList } from '@/features/avatar-picker/utils/avatarUtils';
 import toast from 'react-hot-toast';
 import { useUpdateAvatarMutation } from '@/features/settings/hooks/useUpdateAvatarMutation';
 import { useEffect, useState } from 'react';
 import { useAuthStore } from '@/features/auth/store/useAuthStore';
-import tinycolor from 'tinycolor2';
 import { Button } from '@/shared/components/shadcn/button';
+import BackgroundGrid from '@/features/avatar-picker/components/BackgroundGrid';
+import AvatarGrid from '@/features/avatar-picker/components/AvatarGrid';
 
 const avatarBgColors = Object.values(AVATAR_BG_COLOR);
 
@@ -57,7 +55,7 @@ const AvatarsDrawer = ({ showEditButton = true, showConfirmButton }: AvatarsDraw
     }
 
     try {
-      await updateAvatar({ avatar: selectedAvatarId, backgroundColor: selectedBgColor });
+      updateAvatar({ avatar: selectedAvatarId, backgroundColor: selectedBgColor });
       closeDrawer();
       toast.success('아바타가 성공적으로 업데이트되었습니다!');
     } catch (error) {
@@ -90,102 +88,19 @@ const AvatarsDrawer = ({ showEditButton = true, showConfirmButton }: AvatarsDraw
         </DrawerHeader>
 
         {/* 배경색 그리드 */}
-        <div className="flex items-center px-6 py-6 border-b border-gray-100 mb-5">
-          <div className="flex justify-center gap-5 flex-wrap mx-auto">
-            {avatarBgColors.map(({ token, hex }) => (
-              <button
-                key={hex}
-                onClick={() => setBgColor(hex)}
-                className={cn(
-                  'relative w-12 h-12 rounded-full transition-all duration-200 hover:scale-110 focus:scale-120 hover:shadow-sm cursor-pointer',
-                )}
-                style={{
-                  backgroundColor: token,
-                }}
-                aria-label={`색상 ${hex}`}
-              >
-                {selectedBgColor === hex && (
-                  <div className="absolute inset-0 flex items-center justify-center">
-                    <div className="bg-white rounded-full p-1 shadow-lg">
-                      <Check className="text-gray-800 w-4 h-4" strokeWidth={3} />
-                    </div>
-                  </div>
-                )}
-              </button>
-            ))}
-          </div>
-        </div>
-
+        <BackgroundGrid
+          avatarBgColors={avatarBgColors}
+          setBgColor={setBgColor}
+          selectedBgColor={selectedBgColor}
+        />
         {/* 아바타 그리드 */}
-        <div className="px-24 py-6 max-h-96 overflow-y-auto">
-          <div role="listbox" aria-label="아바타 목록" className="grid grid-cols-4 gap-6">
-            {avatarList.map((avatarUrl, index) => {
-              const isSelected = selectedAvatarId === String(index);
-              const isHovered = hoveredIndex === index;
-
-              return (
-                <div key={index} className="flex justify-center">
-                  <button
-                    type="button"
-                    role="option"
-                    aria-selected={isSelected}
-                    tabIndex={0}
-                    className="group relative cursor-pointer outline-none"
-                    onClick={() => setAvatarId(String(index))}
-                    onMouseEnter={() => setHoveredIndex(index)}
-                    onMouseLeave={() => setHoveredIndex(null)}
-                  >
-                    {isSelected && (
-                      <div
-                        className="absolute inset-0 rounded-full blur-xl opacity-50 scale-110 transition-all"
-                        style={{
-                          background: `radial-gradient(circle, ${selectedBgColor} 0%, transparent 30%)`,
-                        }}
-                      />
-                    )}
-
-                    <Avatar
-                      className={cn(
-                        'w-30 h-30 border-4 transition-all p-1 duration-300 relative flex items-center justify-center',
-                        isSelected
-                          ? 'shadow-lg scale-105'
-                          : 'border-gray-200 group-hover:shadow-md group-hover:scale-105',
-                      )}
-                      style={{
-                        backgroundColor:
-                          isSelected || isHovered ? selectedBgColor || '#f3f4f6' : '#ffffff',
-                        borderColor: isSelected
-                          ? tinycolor(selectedBgColor || '#f3f4f6')
-                              .darken(3)
-                              .toString()
-                          : '#ffffff',
-                      }}
-                    >
-                      <AvatarImage
-                        src={avatarUrl}
-                        alt={`Avatar ${index + 1}`}
-                        className="w-24 h-24"
-                      />
-                      <AvatarFallback
-                        style={{
-                          backgroundColor: selectedBgColor || '#f3f4f6',
-                        }}
-                      >
-                        <User size={24} />
-                      </AvatarFallback>
-                    </Avatar>
-
-                    {isSelected && (
-                      <div className="absolute top-1 -right-0 bg-boost-blue rounded-full p-1.5 shadow-lg">
-                        <Check size={12} className="text-white" />
-                      </div>
-                    )}
-                  </button>
-                </div>
-              );
-            })}
-          </div>
-        </div>
+        <AvatarGrid
+          selectedAvatarId={selectedAvatarId}
+          hoveredIndex={hoveredIndex}
+          setAvatarId={setAvatarId}
+          setHoveredIndex={setHoveredIndex}
+          selectedBgColor={selectedBgColor}
+        />
 
         {showConfirmButton && selectedAvatarId && selectedBgColor && (
           <div className="absolute top-2 left-0 right-2 flex justify-end px-6 py-4 bg-transparent">
